+++ packages/core/src/ClassNameAdapter.ts
@@ -1,11 +1,7 @@
-import Aesthetic from './Aesthetic';
+import Adapter from './Adapter';
 import { ClassName } from './types';
 
-export default class ClassNameAesthetic<Theme extends object> extends Aesthetic<
-  Theme,
-  {},
-  ClassName
-> {
+export default class ClassNameAdapter extends Adapter<{}, ClassName> {
   transformToClassName(styles: unknown[]): ClassName {
     return styles.filter(style => style && typeof style === 'string').join(' ');
   }