+++ packages/core/src/index.ts
@@ -3,8 +3,10 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
+import instance from './instance';
+import Adapter from './Adapter';
 import Aesthetic from './Aesthetic';
-import ClassNameAesthetic from './ClassNameAesthetic';
+import ClassNameAdapter from './ClassNameAdapter';
 import UnifiedSyntax from './UnifiedSyntax';
 import Ruleset from './Ruleset';
 import Sheet from './Sheet';
@@ -12,6 +14,6 @@ import Sheet from './Sheet';
 export * from './constants';
 export * from './types';
 
-export { ClassNameAesthetic, UnifiedSyntax, Ruleset, Sheet };
+export { Adapter, Aesthetic, ClassNameAdapter, UnifiedSyntax, Ruleset, Sheet };
 
-export default Aesthetic;
+export default instance;