+++ packages/react/src/types.ts
@@ -2,18 +2,18 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 import React from 'react';
-import Aesthetic, {
+import {
   ClassNameTransformer,
+  CompiledStyleSheet,
   Direction,
-  SheetMap,
   StyleName,
-  StyleSheetDefinition,
+  StyleSheetFactory,
   ThemeName,
+  ThemeSheet,
 } from 'aesthetic';
 import { Omit } from 'utility-types';
 
 export interface DirectionProviderProps {
-  aesthetic: Aesthetic<any, any, any>;
   children: NonNullable<React.ReactNode>;
   dir?: Exclude<Direction, 'neutral'>;
   inline?: boolean;
@@ -21,7 +21,6 @@ export interface DirectionProviderProps {
 }
 
 export interface ThemeProviderProps {
-  aesthetic: Aesthetic<any, any, any>;
   children: NonNullable<React.ReactNode>;
   name?: ThemeName;
   propagate?: boolean;
@@ -41,7 +40,7 @@ export interface WithThemeWrapperProps {
   wrappedRef?: React.Ref<any>;
 }
 
-export interface WithThemeWrappedProps<Theme> {
+export interface WithThemeWrappedProps<Theme = ThemeSheet> {
   /** The ref passed by the `wrappedRef` prop. Provided by `withTheme`. */
   ref?: React.Ref<any>;
   /** The theme object. Provided by `withTheme`. */
@@ -58,17 +57,13 @@ export interface WithStylesWrapperProps {
   wrappedRef?: React.Ref<any>;
 }
 
-export interface WithStylesWrappedProps<
-  Theme,
-  NativeBlock extends object,
-  ParsedBlock extends object | string = NativeBlock
-> {
+export interface WithStylesWrappedProps<Theme = ThemeSheet> {
   /** Utility function to transform parsed styles into CSS class names. Provided by `withStyles`. */
-  cx: ClassNameTransformer<NativeBlock, ParsedBlock>;
+  cx: ClassNameTransformer;
   /** The ref passed by the `wrappedRef` prop. Provided by `withStyles`. */
   ref?: React.Ref<any>;
   /** The parsed component style sheet in which rulesets can be transformed to class names. Provided by `withStyles`. */
-  styles: SheetMap<ParsedBlock>;
+  styles: CompiledStyleSheet;
   /** The theme object when `passThemeProp` is true. Provided by `withStyles`. */
   theme?: Theme;
 }
@@ -90,7 +85,7 @@ export interface WithStylesOptions {
 
 // Name is based on react-docgen-typescript:
 // https://github.com/styleguidist/react-docgen-typescript/blob/master/src/parser.ts#L850
-export interface StyledComponent<Theme, Props> extends React.NamedExoticComponent<Props> {
+export interface StyledComponent<Props> extends React.NamedExoticComponent<Props> {
   displayName: string;
 
   styleName: StyleName;
@@ -98,7 +93,7 @@ export interface StyledComponent<Theme, Props> extends React.NamedExoticComponen
   WrappedComponent: React.ComponentType<any>;
 
   extendStyles<T>(
-    styleSheet: StyleSheetDefinition<Theme, T>,
+    styleSheet: StyleSheetFactory<ThemeSheet, T>,
     extendOptions?: Omit<WithStylesOptions, 'extendFrom'>,
-  ): StyledComponent<Theme, Props>;
+  ): StyledComponent<Props>;
 }