+++ packages/adapter-typestyle/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 4.0.0
+
+#### 💥 Breaking
+
+- Rewritten to support the
+  [core 5.0 changes](https://github.com/milesj/aesthetic/blob/master/packages/aesthetic/CHANGELOG.md).
+- Adapter no longer extends `Aesthetic` and does not accept options in the constructor.
+
 ### 3.0.1 - 2019-09-15
 
 #### 📦 Dependencies